lg-agent-docker/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "JodyHuntATX"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain-openai (>=0.3.0,<0.4.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "mysql-connector-python (>=9.1.0,<10.0.0)",
    "langgraph (>=0.2.62,<0.3.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
